// src/redux/action-types/Host.types.ts
export enum HostActionTypes {
  SET_HOSTS = "SET_HOSTS",
  SET_DRAFT_HOSTS = "SET_DRAFT_HOSTS",
  UPDATE_HOST_PENDING = "UPDATE_HOST_PENDING",
  UPDATE_HOST_SUCCESS = "UPDATE_HOST_SUCCESS",
  UPDATE_HOST_FAILED = "UPDATE_HOST_FAILED",
  UPDATE_DRAFT_HOST_PENDING = "UPDATE_DRAFT_HOST_PENDING",
  UPDATE_DRAFT_HOST_SUCCESS = "UPDATE_DRAFT_HOST_SUCCESS",
  UPDATE_DRAFT_HOST_FAILED = "UPDATE_DRAFT_HOST_FAILED",
  TOOGLE_OPEN_ALERT = "TOOGLE_OPEN_ALERT",
  ADD_VULNERBILITY_PENDING = "ADD_VULNERBILITY_PENDING",
  ADD_VULNERBILITY_SUCCESS = "ADD_VULNERBILITY_SUCCESS",
  ADD_VULNERBILITY_FAILED = "ADD_VULNERBILITY_FAILED",
  REMOVE_VULNERBILITY_PENDING = "REMOVE_VULNERBILITY_PENDING",
  REMOVE_VULNERBILITY_SUCCESS = "REMOVE_VULNERBILITY_SUCCESS",
  REMOVE_VULNERBILITY_FAILED = "REMOVE_VULNERBILITY_FAILED",
  ADD_SERVICE_PENDING = "ADD_SERVICE_PENDING",
  ADD_SERVICE_SUCCESS = "ADD_SERVICE_SUCCESS",
  ADD_SERVICE_FAILED = "ADD_SERVICE_FAILED",
  REMOVE_SERVICE_PENDING = "REMOVE_SERVICE_PENDING",
  REMOVE_SERVICE_SUCCESS = "REMOVE_SERVICE_SUCCESS",
  REMOVE_SERVICE_FAILED = "REMOVE_SERVICE_FAILED",
}